% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorTest.R
\name{e.cortest}
\alias{e.cortest}
\title{ecortest}
\usage{
e.cortest(x, y, stat = "diff", histobins = 500, histoLeft = NA,
  histoRight = NA, safeSecs = 30, safeSets = 1e+10, plot = T,
  detail = 2, nreps = 20, rseed = 0, B = 0)
}
\arguments{
\item{x}{array of x values to compare with y}

\item{y}{array with same length as x. y[1] is compared with x[1] etc.}

\item{stat}{currently ignored}

\item{histobins}{number of bins for distribution}

\item{histoLeft}{left end of plotted distribution. Leave as NA to have it set automatically.}

\item{histoRight}{right end of distribution. NA means set automatically}

\item{safeSecs}{Stop the calculations after this many seconds}

\item{safeSets}{Do not attempt full enumeration if there are more than this many subsets}

\item{plot}{Display distribution histogram if TRUE.}

\item{detail}{How much detail to print after completion of test}

\item{nreps}{Trials in monte carlo version are performed in groups of size nreps. Optimum seems to be around 20}

\item{rseed}{seeds random numbers for monte carlo. Default is current time, but can be set programmatically if more than run per second is being done.}

\item{B}{The number of random trials to use. Set to zero to use full enumeration.}
}
\value{
Returns a list which includes the following
\item{x}{vector containing one member of each pair}
\item{y}{vector same length as x but with other member of each pair} 
\item{n}{the number of data pairs}
\item{nTodo}{the number of monte carlo trials requested. Not used in full enumeration}
\item{statLeft}{correlations less than this are in left tail}
\item{statRight}{correlations greater than this are in right tail}
\item{histobins}{number of bins used in distribution histogram}
\item{histoLeft, histoRight}{boundaries of plotted histogram}
\item{safeSecs}{process is aborted if it exceeds this many seconds}
\item{safeSets}{process is aborted if it exceeds this many trials}
\item{nlower, nupper}{the actual numbers of subsets found in the left and right tail respectively}
\item{histoData}{observed number of subsets in each bin. length = histobins}
\item{subsetsExamined}{the actual, as opposed to requested, number of subsets tested}
\item{timespent}{how long was spent in C code doing the test}
\item{statObs}{the correlation coefficient corresponding to the observed permutation -- will be equal to either statLeft or statRight}
\item{statTail}{negative of statObs}
\item{P2t}{two-tail P value}
\item{P1t}{one-tail P value -- direction depends on the direction of observed data}
\item{testname}{"Pearson Correlation"}
}
\description{
Compare x versus y  
Uses efficient algorithms for either complete enumeration or monte carlo test. 
Resulting distribution is displayed graphically. 
Test statistic is Pearson correlation
Trials consist of permutations
}
\details{

}
\examples{
x <- rnorm(10)
y <- x + rnorm(10, 5)
result <- e.cortest(x, y) # performs full enumeration and plots result
result <- e.cortest(x, y, B=1e6) # tests one million random permutations

}
