% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epairTest.R
\name{e.pairtest}
\alias{e.pairtest}
\title{epairtest}
\usage{
e.pairtest(x, y, stat = c("diff"), histobins = 500, histoLeft = NA,
  histoRight = NA, safeSecs = 30, safeSets = 1e+10, plot = T,
  detail = 2, rseed = 0, B = 0)
}
\arguments{
\item{x}{array of x values to compare with y}

\item{y}{array with same length as x. y[1] is compared with x[1] etc.}

\item{stat}{currently ignored}

\item{histobins}{number of bins for distribution}

\item{histoLeft}{left end of plotted distribution. Leave as NA to have it set automatically.}

\item{histoRight}{right end of distribution. NA means set automatically}

\item{safeSecs}{Stop the calculations after this many seconds}

\item{safeSets}{Do not attempt full enumeration if there are more than this many subsets}

\item{plot}{Display distribution histogram if TRUE.}

\item{detail}{How much detail to print after completion of test (currently 0 or 1)}

\item{rseed}{supplements current time to seed random numbers. Can be ignored for full enumeration or if no more than 1 run per second.}

\item{B}{The number of random trials to use. Set to zero to use full enumeration.}
}
\value{
Returns a list which includes the following
\item{x}{vector of differences (\code{x-y})}
\item{n}{the number of pairs = length of x or y}
\item{nTodo}{the number of monte carlo trials requested. Not used in full enumeration}
\item{statLeft}{mimimum possible value of the sum of differences statistic}
\item{statRight}{maximum possible sum of differences. \code{statRight = -statLeft}}
\item{histobins}{number of bins used in histogram}
\item{histoLeft, histoRight}{boundaries of plotted histogram}
\item{safeSecs}{process aborted after this many seconds}
\item{nlower, nupper}{the numbers of trials falling into the left and right tails}
\item{histoData}{found numbers of trials in each bin of histogram}
\item{subsetsExamined}{The actual number of combinations examined, as opposed to the requested number}
\item{timespent}{the time in seconds spent in the C routine}
\item{statObs}{observed sum of differences}
\item{statTail}{negative of \code{statObs}}
\item{P2t}{two-tail P value}
\item{P1t}{one-tail P value. The tail is always chosed as the one in the same direction as observed}
\item{testname}{"paired:    mean(x-y)"}
}
\description{
Compare x versus y when values are paired. 
Uses efficient algorithms for either complete enumeration or monte carlo test. 
Resulting distribution is displayed graphically. 
Test statistic is sum of x values or the mean of the x values
Trials consist of pairings
}
\details{

}
\examples{
x <- rnorm(10)
y <- rnorm(10)
result <- e.pairtest(x, y) # performs full enumeration and plots results
result <- e.pairtest(x, y, B=1e8) # same test, but uses 100 million random trials instead of full enumeration

}
