% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esumtest.R
\name{e.sumtest}
\alias{e.sumtest}
\title{esumtest}
\usage{
e.sumtest(x, y, stat = c("diff", "mean"), histobins = 500,
  histoLeft = NA, histoRight = NA, safeSecs = 30, safeSets = 1e+10,
  plot = T, detail = 2, nreps = 20, rseed = 0, B = 0)
}
\arguments{
\item{x}{array of x values to compare with y}

\item{y}{array of y values}

\item{stat}{Set to \dQuote{diff} to use the difference between the mean of x and mean of y. Set to \dQuote{mean} to use just the x mean as test statistic. P value is not affected.}

\item{histobins}{number of bins for distribution}

\item{histoLeft}{left end of plotted distribution. Leave as NA to have it set automatically.}

\item{histoRight}{right end of distribution. NA means set automatically}

\item{safeSecs}{Stop the calculations after this many seconds}

\item{safeSets}{Do not attempt full enumeration if there are more than this many subsets}

\item{plot}{Display distribution histogram if TRUE.}

\item{detail}{How much detail to print after completion of test}

\item{B}{The number of random trials to use. Set to zero to use full enumeration.}
}
\value{
Returns a list which includes the following
\item{x}{all values, including original x and y}
\item{n}{the total number of values}
\item{k}{the number of values in first set}
\item{B}{the number of requested random trials, or zero if full enumeration}
\item{statLeft, statRight}{the cutoffs for computing P values, one- or two-tail}
\item{histobins}{number of bins in distribution histogram}
\item{histoLeft, histoRight}{boundaries of histogram}
\item{safeSecs, safeSets}{time and subset limits to prevent getting stuck in C routine longer than intended}
\item{nlower, nupper}{the actual numbers of subsets found in the upper and lower tails}
\item{histoData}{observed number of subsets in each bin}
\item{subsetsExamined}{the actual, as opposed to requested, number of subsets tested}
\item{timespent}{the time in seconds spent in the C routine}
\item{P2t}{two-tail P value}
\item{P1t}{one-tail P value -- direction depends on the direction of observed data}
\item{testname}{which test statistic was used for histogram plot}
}
\description{
Compare sets of values, x and y, for mean differences with exact test. 
Uses efficient algorithms for either complete enumeration or monte carlo test. 
Resulting distribution is displayed graphically.
}
\details{
This function calls one of two C routines optimized for speed. One is "esumTest" to perform the full enumeration of all k-subsets of the pooled x and y values. 
The other, "rsumTest", selects random subsets.
The distribution of the test statistic is displayed as a histogram.
Subsets different in the same direction but at least as extreme as observed are shown in red.
Those in the opposite tail are colored blue
An orange curve shows the asymptotic t distribution. Differences between orange curve and true distribution provides evidence for usefulness of exact test.
}
\examples{
x <- rnorm(10, 5)
y <- rnorm(20, 7)
result <- e.sumtest(x,y) #performs full enumeration
result <- e.sumtest(x,y, B=1e6) #tests one million random subsets

}
